==> 1. Crawling/03_youtube_crawling.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.keys import Keys

import requests
from bs4 import BeautifulSoup

import time

import pandas as pd

from pytube import YouTube
import re
from datetime import datetime

import sqlite3

def crawling_youtuber_video_link(path, youtuber, num_end=10):
    """해당 유튜버의 영상 링크들을 가져오기
    
    path : 크롬드라이버 주소
    youtuber : 유튜버의 홈 링크
    num_end : 동영상 탭에서 end 키를 누르는 횟수 => 비디오 링크를 얼마나 가져올 것인가
    
    가져오는 변수를 좀 추가해야되나??
    여기서 더 안가져오고 동영상 링크 들어가면 거기서 많이 가져오면 될 것 같기도 하고
    """
    driver = webdriver.Chrome(path)
    driver.maximize_window() # 창 크기 최대화
    
    driver.get(youtuber) # 해당 링크로 이동
    time.sleep(5)
    driver.find_element_by_xpath('//*[@id="tabsContent"]/paper-tab[2]/div').click() # 버튼 클릭
    
    # end 키를 눌러 페이지 맨 밑으로 이동
    for i in range(num_end):
        driver.find_element_by_tag_name('body').send_keys(Keys.END)
        time.sleep(5)
        
    dom = driver.page_source
    soup = BeautifulSoup(dom, 'html.parser')
    link_source = soup.select("div#details") # 영상들이 있는 부분의 소스 추출
    
    video_link_list = []
    
    for link in link_source:
        title = link.find('a')['title']
        href = 'https://www.youtube.com' + link.find('a')['href']
        video_link_list.append([datetime.now().strftime("%Y/%m/%d %H:%M:%S"), title, href])
    
    driver.close()
    
    youtuber_video_link = pd.DataFrame(video_link_list, columns=['datetime', 'title', 'video_link'])
    
    return youtuber_video_link

def get_video_info(video_link):
    '''pytube를 이용하여 동영상의 제목, 링크, 영상 길이(초), 평점, 썸네일 링크, 조회수, 영상 설명, 자동 생성 자막을 가져오는 함수
    
    video_link : 영상 url
    '''
    
    crawling_time = datetime.now().strftime("%Y/%m/%d %H:%M:%S")
    try:
        # 영상 가져오기
        source = YouTube(video_link)

        # 영상 정보
        title = source.title
        length = source.length
        rating  = source.rating
        thumbnail_url = source.thumbnail_url
        views = source.views
        description = source.description

        # 자동 생성 자막 불러오기 및 전처리
        if 'ko' in source.captions.keys():
            ko_caption = source.captions.get('ko')
            ko_caption_convert_to_srt = ko_caption.generate_srt_captions()

            rule1 = r'\d{1,}\n\d{2}:\d{2}:\d{2}\,\d{3} --> \d{2}:\d{2}:\d{2}\,\d{3}\n'
            rule2 = r'\n'
            rule3 = r'\[\w{1,}\]'
            rule4 = '니다'

            ko_caption_convert_to_srt_ = re.sub(rule4, '니다.',
                                                re.sub(rule3, '',
                                                       re.sub(rule2, '', 
                                                              re.sub(rule1, ' ', ko_caption_convert_to_srt))))
        else:
            ko_caption_convert_to_srt_ = 'No caption in video'

        video_info = [crawling_time, title, video_link, length, rating, thumbnail_url, views, description, ko_caption_convert_to_srt_]
    
    except:
        video_info = [None, None, None, None, None, None, None, None, None]
    
    return video_info

def get_video_infos(video_links):
    '''복수의 video_link에 대해 영상 정보 가져오기
    
    video_links : sequence of video_link
    '''
    video_info_list = []
    
    for video_link in video_links:
        video_info = get_video_info(video_link)
        video_info_list.append(video_info)
        
    video_info_df = pd.DataFrame(video_info_list,
                                 columns=['crawling_time', 'title', 'video_link', 'length', 'rating',
                                          'thumbnail_url', 'views', 'description', 'ko_caption'])
    
    return video_info_df

### 유튜버의 영상 링크들 가져오기

selenium 안써야 일관성 있게 가져올 수 있을거같은데... ㅠㅠ 찾아보다가 잘 모르겠어서 일단 다른거부터 했습니다

ytn_video_list = crawling_youtuber_video_link('chromedriver.exe', 'https://www.youtube.com/user/ytnnews24/featured', num_end=10)

ytn_video_list

### 해당 링크의 영상 정보 (자막 등)

ytn_video_infos = get_video_infos(ytn_video_list['video_link'][0:100])

ytn_video_infos

print(ytn_video_infos.loc[0, 'video_link'])
print(ytn_video_infos.loc[0, 'ko_caption'])

print(ytn_video_infos.loc[50, 'video_link'])
print(ytn_video_infos.loc[50, 'ko_caption'])

print(ytn_video_infos.loc[90, 'video_link'])
print(ytn_video_infos.loc[90, 'ko_caption'])

### db 생성해서 table로 넣기

conn = sqlite3.connect('./youtube_data.db')
cur = conn.cursor()

ytn_video_list.to_sql('ytn_video_list', conn)

ytn_video_infos.to_sql('ytn_video_info', conn)

cur.close()
conn.close()

어떻게 하기는 했는데...ㅠㅠㅠㅠ 뭔가 너무 못한거같아서 ㅠㅠㅠㅠ 완성도가 떨어져서... 정말 죄송합니다 ㅠㅠ

일단 보면서 느낀건 영상 자동 생성 자막에 부정확한 부분이 많아서 그대로 사용하면 되는건가? 라는 생각이 들었고

그래서 주요 뉴스 사이트 들어가보니까 정보 제공 현황이 아래와 같았습니다.

ytn : 영상 설명에 내용 O, 원문 보기 링크 O

kbs : 영상 설명에 내용 O, 원문 보기 링크 X

mbc : 영상 설명에 내용 X, 원문 보기 링크 O

sbs : 영상 설명에 내용 X, 원문 보기 링크 O

연합뉴스 : 영상 설명에 내용 O, 원문 보기 링크 X

그래서 해당 채널별로 따로 코드를 짜서 크롤링을 해야되나?? 라는 생각이 들었고

만약에 케이스를 나눠서 하면 원문 링크가 있으면 거기 들어가서 가져오는 식으로 짜면 될 것 같은데

원문 링크가 없고 영상 설명에만 있으면 그걸 가져온다 했을 때 영상 길이가 길면 영상 설명에 있는 대사를 중략에서 올리더라고요

그래서 영상 설명에 있는 내용을 가져오는 방식으로 하려면 영상 길이에 제약을 둬야 할 것 같네요 (ex. 5분 이하로?)

그리고 네이버 뉴스, 다음 뉴스 요약본 몇 개 보면서 느낀 건

기사 전체를 뜯어서 핵심 내용 파악해서 문장을 재구성해서 요약하는게 아니고

기사 전체의 첫 번째 문장 + 아래의 내용에서 중요하다고 판단되는 문장 2~3개를 그대로 가져오는 방식이더라고요.

저희가 의도하는 것도 그런 방식?으로 진행되는건지 이야기해보면 좋을 것 같다고 생각했어요.
