==> 1. Crawling/05_youtube_crawling_api_pytube.ipynb <==
# 함수 설계

### 1. 비디오 링크 가져오기

1. 입력 인자

유튜버 아이디 입력 (youtuber_id=ytnnews24) + 시작 날짜와 종료 날짜 입력 (start_date=2020-07-01, end_date=2020-08-01)

2. 데이터 불러오는 방법

youtubeAPI 이용하여 그 날짜에 해당하는 동영상 video_url들을 추출

영살 길이가 10분 이하인 동영상들만 가져옴

3. 결과

columns = ['published_time', 'video_id', 'video_url'] 인 DataFrame

### 2. 비디오 링크에 접속하여 영상 정보 가져오기

1. 입력 인자

youtubeAPI 이용하여 만든 DataFrame

2. 데이터 불러오는 방법

pytube : 영상 정보

youtube_transcript_api : 자동 생성 자막을 {내용, 시간, 지속시간} 형태인 dictionary

#### => dictionary라서 sql에 안들어가길래 csv로 따로 저장해놨어요

자동 생성 스크립트의 품질이 별로니까 다음과 같이 가져와보자. 만약 둘 다 있으면 상황 보고 편한걸로 가져오기.

방법 1. 영상 설명에 기사 전문이 있다 => 정규표현식을 이용하여 필요한 정보만 깔끔하게 걸러내기

방법 2. 영상 설명에 기사 전문이 없고 원문 링크만 있다 => 정규표현식으로 원문 링크 가져와서 거기 접속하여 기사 전문 깔끔하게 만들어 가져오기


3. 결과

columns = ['published_time', 'video_id', 'video_url', 'crawling_time', 'author', 'title', 'length', 'rating', 'views',
           'thumbnail_url', 'captions', 'description'] 인 DataFrame


# 뉴스 유튜브 채널 자료 제공 현황

이렇게 8개의 채널만 가져와도 충분할 것이라고 생각합니다!

## 영상 설명에 내용 있는 언론사

1. ytn : 영상 설명에 내용 O, 원문 보기 링크 O, 링크 : https://www.youtube.com/user/ytnnews24/

채널 id : UChlgI3UHCOnwUGzWzbJ3H5w

2. 연합뉴스 : 영상 설명에 내용 O, 원문 보기 링크 X, 링크 : https://www.youtube.com/channel/UCTHCOPwqNfZ0uiKOvFyhGwg

채널 id : UCTHCOPwqNfZ0uiKOvFyhGwg

3. 채널A : 영상 설명에 내용 O, 원문 보기 링크 X, 링크 : https://www.youtube.com/user/tvchanews

채널 id : UCfq4V1DAuaojnr2ryvWNysw

4. MBN : 영상 설명에 내용 O, 원문 보기 링크 X, 링크 : https://www.youtube.com/user/mbn

채널 id : UCG9aFJTZ-lMCHAiO1KJsirg

## 영상 설명에 내용 없고 원문 보기 링크가 있는 언론사

원문 보기 링크가 없는 동영상은 삭제하는 방식으로 가야겠다

5. kbs : 영상 설명에 내용 X, 원문 보기 링크 O, 링크 : https://www.youtube.com/user/NewsKBS

원문 보기 링크를 들어가도 해당 기사 페이지가 안나오고 kbs news 메인 페이지가 나옴!

=> 메인 언론사이지만 제외하는게 속 편할 것 같다는 생각을 했어요.

채널 id : UCcQTRi69dsVYHN3exePtZ1A

6. mbc : 영상 설명에 내용 X, 원문 보기 링크 O, 링크 : https://www.youtube.com/user/MBCnews

채널 id : UCF4Wxdo3inmxP-Y59wXDsFw

원문 보기 링크 예시

5분 37초 : https://imnews.imbc.com/replay/2020/nwtoday/article/5865528_32531.html

2분 40초 : https://imnews.imbc.com/replay/2020/nwtoday/article/5865560_32531.html

7. sbs : 영상 설명에 내용 X, 원문 보기 링크 O, 링크 : https://www.youtube.com/user/sbsnews8

채널 id : UCkinYTS9IHqOEwR1Sze2JTw

원문 보기 링크 예시

2분 24초 : https://news.sbs.co.kr/news/endPage.do?news_id=N1005918079&plink=LINK&cooper=YOUTUBE

3분 47초 : https://news.sbs.co.kr/news/endPage.do?news_id=N1005917222&plink=LINK&cooper=YOUTUBE

4분 몇초짜리 연애기사는 시간이 짧았음에도 불구하고 원문 기사가 없었음

5분 32초 : https://news.sbs.co.kr/news/endPage.do?news_id=N1005915372&plink=LINK&cooper=YOUTUBE

16분 4초 : https://news.sbs.co.kr/news/endPage.do?news_id=N1005916633&plink=LINK&cooper=YOUTUBE

8. jtbc : 영상 설명에 내용 X, 원문 보기 링크 O, 링크 : https://www.youtube.com/user/JTBC10news

채널 id : UCsU-I-vHLiaMfV_ceaYz5rQ

원문 보기 링크 예시

2분 14초 : 없음

6분 9초 : https://www.youtube.com/watch?v=LqzWyUxN8yU 원문 보기 링크는 없는데 영상 설명이 좀 자세한 편

20분 31초 : 없음

6분 12초 : http://news.jtbc.joins.com/html/210/NB11963210.html

이게 있는게 있고 없는게 있네

데이터 수집 단계에서 정제를 하고 싶었는데 막막해서... 일단 8월 1일 ~ 8월 7일에 대한 영상들을 수집

정제 필요!

# 시간 관리 및 데이터 정제
from datetime import datetime
import time
import re

# 데이터 관리
import pandas as pd
import sqlite3

# 유튜브 정보 추출
from apiclient.discovery import build
from pytube import YouTube
from youtube_transcript_api import YouTubeTranscriptApi

# 웹 정보 수집
import requests
from bs4 import BeautifulSoup


def transform_duration(time):
    '''
    유튜브 동영상 duration이 지저분해서 깔끔하게 바꾸는 함수
    
    time : video duration ex. 'PT3H14M16S'
    '''
    time_list = re.split(r'H|M|S', re.sub(r'PT', '', time))
    
    if len(time_list) == 4:
        time_as_second = 3600*int(time_list[0]) + 60*int(time_list[1]) + int(time_list[2])
    elif len(time_list) == 3:
        time_as_second = 60*int(time_list[0]) + int(time_list[1])
    elif len(time_list) == 2:
        time_as_second = int(time_list[0])
    
    return time_as_second

def get_channel_video_id(channel_id, api_key, start_time, end_time):
    '''
    유튜버의 영상 중 해당 기간 내에 업로드 된 영상들을 가져오는 함수
    
    channel_id : 유튜버의 채널 id
    
    api_key : YoutubeAPI에서 발급받은 api key
    
    start_time : 영상을 가져올 시작 시간 ex. '2020-08-01 00:00:00'
    
    end_time : 영상을 가져올 종료 시간 ex. '2020-08-04 00:00:00'
    '''
    # youtube resource에 접근하기 위한 객체 생성
    youtube = build('youtube', 'v3', developerKey=api_key)
    
    # 해당 채널에서 '업로드 된 동영상' 재생 목록의 id 값을 가져오기
    channel_res = youtube.channels().list(id=channel_id,
                                          part='contentDetails').execute()
    playlist_id = channel_res['items'][0]['contentDetails']['relatedPlaylists']['uploads']
    
    # 자료 저장과 종료 조건 체크를 위한 객체 생성
    videos = []
    video_info_list = []
    base_url = 'https://www.youtube.com/watch?v='
    next_page_token = None
    
    # 데이터 수집
    while True:
        # 재생 목록에 대해 영상 list를 추출, 가장 최신의 동영상부터 순차적으로 가져옴
        video_res = youtube.playlistItems().list(playlistId=playlist_id,
                                                 part='contentDetails',
                                                 maxResults=50, 
                                                 pageToken=next_page_token).execute()
        # 저장 조건 체크 및 자료 저장
        for item in video_res['items']:
            try:
                temp_date = re.sub('Z', '',re.sub('T', ' ', item['contentDetails']['videoPublishedAt']))

                if start_time <= temp_date <= end_time:
                    videos.append(item)

                    published_time = temp_date
                    video_id = item['contentDetails']['videoId']
                    video_url = base_url + item['contentDetails']['videoId']

                    # 영상 길이가 10분 이하일때 저장 : 너무 길면 pytube.YouTube로 가져오는데 너무 오래 걸림
                    res_content = youtube.videos().list(id=video_id, part='contentDetails').execute()
                    duration = transform_duration(res_content['items'][0]['contentDetails']['duration'])

                    if duration <= 600:
                        video_info_list.append([published_time, video_id, video_url])
            
            except:
                continue
        
        # 종료 조건을 위한 값 생성 및 조건 체크
        next_page_token = video_res.get('nextPageToken')
        criteria_date = re.sub('Z', '',
                               re.sub('T', ' ',
                                      video_res['items'][0]['contentDetails']['videoPublishedAt']))
        
        # 종료 조건 1 : 새로 수집 된 영상의 업로드 시간이 시작 시간보다 앞설 때 종료
        if criteria_date < start_time:
            break
            
        # 종료 조건 2 : 더 이상 수집할 데이터가 없을 때 종료    
        if next_page_token is None:
            break
    
    video_info_df = pd.DataFrame(video_info_list, columns=['published_time', 'video_id', 'video_url'])
    video_info_df.sort_values(by='published_time', inplace=True, ignore_index=True)
    
    return video_info_df

def get_video_info(case):
    '''
    pytube를 이용하여 동영상의 저자, 제목, 링크, 영상 길이(초), 평점, 썸네일 링크, 조회수, 영상 설명, 자동 생성 자막을 가져오는 함수
    영상의 description에 뉴스 영상 스크립트가 있으면 이 함수를 사용
    사용 가능 언론사 : YTN, KBS, 연합뉴스, 채널A, MBN
    
    case : get_channel_video_id의 return DataFrame 중 하나의 case
    '''
    published_time = case['published_time']
    video_id = case['video_id']
    video_url = case['video_url']
    
    try:
        # 영상 가져오기
        source = YouTube(video_url)
        
        # 영상 정보
        crawling_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
        author = source.author
        title = source.title
        length = source.length
        rating= source.rating
        views = source.views
        thumbnail_url= source.thumbnail_url
        
        try:
            captions = YouTubeTranscriptApi.get_transcript(video_id, languages=['ko'])
        except:
            captions = None
        
        description = source.description
        
        video_info = [published_time, video_id, video_url,
                      crawling_time, author, title, length, rating, views, thumbnail_url, captions, description]
    
    # 영상 정보를 가져오는데 실패한 경우
    except: 
        video_info = [published_time, video_id, video_url,
                      None, None, None, None, None, None, None, None, None]
    
    return video_info

def get_video_infos(video_info_df):
    '''
    복수의 video_url에 대해 영상 정보 가져오기
    사용 가능 언론사 : YTN, KBS, 연합뉴스, 채널A, MBN
    
    video_info_df : get_channel_video_id의 return DataFrame
    '''
    video_info_list = []
    row_number = len(video_info_df)
    start = time.time()

    for index in range(row_number):
        case = video_info_df.iloc[index]
        video_info = get_video_info(case)
        video_info_list.append(video_info)
        
        # 진행상황 출력
        if index % 100 == 0:
            during = time.time() - start
            during_min = int(during // 60)
            during_second = round(during % 60)
            
            print('진행 상황 : {} / {} ({}%), {}분 {}초동안 돌아가는 중'\
                  .format(index, row_number, round(100 * index / row_number), during_min, during_second))
    
    # DataFrame으로 저장
    video_info_df = pd.DataFrame(video_info_list,
                                 columns=['published_time', 'video_id', 'video_url',
                                          'crawling_time', 'author', 'title', 'length', 'rating', 'views',
                                          'thumbnail_url', 'captions', 'description'])
    
    return video_info_df

my_api_key = 'AIzaSyAGPpzlcDacxNGCoffrH3eK7sOWOEIMUtg'

### 1. ytn

ytn_video_0801_0807 = get_channel_video_id(channel_id='UChlgI3UHCOnwUGzWzbJ3H5w',
                                      api_key=my_api_key,
                                      start_time='2020-08-01 00:00:00',
                                      end_time='2020-08-08 00:00:00')

ytn_video_info_0801_0807 = get_video_infos(ytn_video_0801_0807)

ytn_video_info_0801_0807.to_csv('ytn_video_info_0801_0807.csv')

conn = sqlite3.connect('./youtube.db')
cur = conn.cursor()

ytn_video_info_0801_0807.drop('captions', axis=1, inplace=True)
ytn_video_info_0801_0807.to_sql('ytn_video_info_0801_0807', conn)
cur.close()
conn.close()

### 2. 연합뉴스

yoenhab_video_0801_0807 = get_channel_video_id(channel_id='UCTHCOPwqNfZ0uiKOvFyhGwg',
                                      api_key=my_api_key,
                                      start_time='2020-08-01 00:00:00',
                                      end_time='2020-08-08 00:00:00')

yoenhab_video_info_0801_0807 = get_video_infos(yoenhab_video_0801_0807)

yoenhab_video_info_0801_0807.to_csv('yoenhab_video_info_0801_0807.csv')

conn = sqlite3.connect('./youtube.db')
cur = conn.cursor()

yoenhab_video_info_0801_0807.drop('captions', axis=1, inplace=True)
yoenhab_video_info_0801_0807.to_sql('yoenhab_video_info_0801_0807', conn)
cur.close()
conn.close()

### 3. 채널 a

channel_a_video_0801_0807 = get_channel_video_id(channel_id='UCfq4V1DAuaojnr2ryvWNysw',
                                      api_key=my_api_key,
                                      start_time='2020-08-01 00:00:00',
                                      end_time='2020-08-08 00:00:00')

channel_a_video_info_0801_0807 = get_video_infos(channel_a_video_0801_0807)

channel_a_video_info_0801_0807.to_csv('channel_a_video_info_0801_0807.csv')

conn = sqlite3.connect('./youtube.db')
cur = conn.cursor()

channel_a_video_info_0801_0807.drop('captions', axis=1, inplace=True)
channel_a_video_info_0801_0807.to_sql('channel_a_video_info_0801_0807', conn)
cur.close()
conn.close()

### 4. mbn

mbn_video_0801_0807 = get_channel_video_id(channel_id='UCG9aFJTZ-lMCHAiO1KJsirg',
                                      api_key=my_api_key,
                                      start_time='2020-08-01 00:00:00',
                                      end_time='2020-08-08 00:00:00')

mbn_video_info_0801_0807 = get_video_infos(mbn_video_0801_0807)

mbn_video_info_0801_0807.to_csv('mbn_video_info_0801_0807.csv')

conn = sqlite3.connect('./youtube.db')
cur = conn.cursor()

mbn_video_info_0801_0807.drop('captions', axis=1, inplace=True)
mbn_video_info_0801_0807.to_sql('mbn_video_info_0801_0807', conn)
cur.close()
conn.close()

### 5. kbs

kbs_video_0801_0807 = get_channel_video_id(channel_id='UCcQTRi69dsVYHN3exePtZ1A',
                                      api_key=my_api_key,
                                      start_time='2020-08-01 00:00:00',
                                      end_time='2020-08-08 00:00:00')

kbs_video_info_0801_0807 = get_video_infos(kbs_video_0801_0807)

kbs_video_info_0801_0807.to_csv('kbs_video_info_0801_0807.csv')

conn = sqlite3.connect('./youtube.db')
cur = conn.cursor()

kbs_video_info_0801_0807.drop('captions', axis=1, inplace=True)
kbs_video_info_0801_0807.to_sql('kbs_video_info_0801_0807', conn)
cur.close() 
conn.close()

### 6. mbc

mbc_video_0801_0807 = get_channel_video_id(channel_id='UCF4Wxdo3inmxP-Y59wXDsFw',
                                      api_key=my_api_key,
                                      start_time='2020-08-01 00:00:00',
                                      end_time='2020-08-08 00:00:00')

mbc_video_info_0801_0807 = get_video_infos(mbc_video_0801_0807)

mbc_video_info_0801_0807.to_csv('mbc_video_info_0801_0807.csv')

conn = sqlite3.connect('./youtube.db')
cur = conn.cursor()

mbc_video_info_0801_0807.drop('captions', axis=1, inplace=True)
mbc_video_info_0801_0807.to_sql('mbc_video_info_0801_0807', conn)
cur.close()
conn.close()

### 7. sbs

sbs_video_0801_0807 = get_channel_video_id(channel_id='UCkinYTS9IHqOEwR1Sze2JTw',
                                      api_key=my_api_key,
                                      start_time='2020-08-01 00:00:00',
                                      end_time='2020-08-08 00:00:00')

sbs_video_info_0801_0807 = get_video_infos(sbs_video_0801_0807)

sbs_video_info_0801_0807.to_csv('sbs_video_info_0801_0807.csv')

conn = sqlite3.connect('./youtube.db')
cur = conn.cursor()

sbs_video_info_0801_0807.drop('captions', axis=1, inplace=True)
sbs_video_info_0801_0807.to_sql('sbs_video_info_0801_0807', conn)
cur.close()
conn.close()

### 8. jtbc

jtbc_video_0801_0807 = get_channel_video_id(channel_id='UCsU-I-vHLiaMfV_ceaYz5rQ',
                                      api_key=my_api_key,
                                      start_time='2020-08-01 00:00:00',
                                      end_time='2020-08-08 00:00:00')

jtbc_video_info_0801_0807 = get_video_infos(jtbc_video_0801_0807)

jtbc_video_info_0801_0807.to_csv('jtbc_video_info_0801_0807.csv')

conn = sqlite3.connect('./youtube.db')
cur = conn.cursor()

jtbc_video_info_0801_0807.drop('captions', axis=1, inplace=True)
jtbc_video_info_0801_0807.to_sql('jtbc_video_info_0801_0807', conn)
cur.close()
conn.close()

### 데이터 수집 결과

ytn_video_info_0801_0807.shape

yoenhab_video_info_0801_0807.shape

channel_a_video_info_0801_0807.shape

mbn_video_info_0801_0807.shape

kbs_video_info_0801_0807.shape

mbc_video_info_0801_0807.shape

sbs_video_info_0801_0807.shape

jtbc_video_info_0801_0807.shape

data_shape = []
data_shape.append(ytn_video_info_0801_0807.shape)
data_shape.append(yoenhab_video_info_0801_0807.shape)
data_shape.append(channel_a_video_info_0801_0807.shape)
data_shape.append(mbn_video_info_0801_0807.shape)
data_shape.append(kbs_video_info_0801_0807.shape)
data_shape.append(mbc_video_info_0801_0807.shape)
data_shape.append(sbs_video_info_0801_0807.shape)
data_shape.append(jtbc_video_info_0801_0807.shape)

row = [item[0] for item in data_shape]
sum(row)

영상 설명에 내용 있는 언론사 => 정규표현식으로 get_video_info로 가져온 결과의 description cleaning

영상 설명에 내용 없고 원문 보기 링크가 있는 언론사 => description에서 링크 추출

링크에 접속해서 원문 가져오기 => 정규표현식으로 cleaning
